fix status 5

+++ models/STATUS_EQUIPMENT.php
@@ -32,7 +32,7 @@
         }
         public function create_update(){
 
-            $query = 'CREATE OR REPLACE FUNCTION merge_db(key varchar,brand_c varchar,devip_c varchar,subid_c varchar,opid_c varchar,status_c smallint,mode_c smallint,linkstatus_c smallint,palletno_c varchar,mission_c Text,palletsize_c varchar) RETURNS VOID AS
+            $query = 'CREATE OR REPLACE FUNCTION merge_db(key varchar,brand_c varchar,devip_c varchar,subid_c varchar,opid_c varchar,status_c integer,mode_c integer,linkstatus_c integer,palletno_c varchar,mission_c Text,palletsize_c varchar) RETURNS VOID AS
             $$
             BEGIN
                     UPDATE '.$this->table.' SET brand =brand_c,devip=devip_c,subid=subid_c,opid=opid_c,status=status_c,mode=mode_c,linkstatus=linkstatus_c,palletno=palletno_c,mission=mission_c,palletsize=palletsize_c  WHERE equipid = key;
